requirements.txt
agno
duckduckgo-search
python-dotenv
anthropic